# docs/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplyFlow Contract Intake",
    "description": "API for uploading contracts, extracting metadata, and storing in Airtable",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://complyflow-production.up.railway.app",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/contracts/upload": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Upload Contract",
        "description": "Upload a contract PDF for processing.\n\nThis endpoint:\n1. Extracts text from the PDF\n2. Uses LLM to extract metadata (parties, dates, terms, etc.)\n3. Computes derived dates (notice deadline, renewal date)\n4. Stores the contract in Airtable with status \"under_review\"\n5. Sends a Slack notification (if configured)",
        "operationId": "upload_contract_contracts_upload_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_contract_contracts_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (bad file, empty, too large)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "LLM service error after retries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "LLM timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{record_id}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get Contract",
        "description": "Get a contract by its Airtable record ID.",
        "operationId": "get_contract_contracts__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractRecord"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Delete Contract",
        "description": "Delete a contract by its Airtable record ID.",
        "operationId": "delete_contract_contracts__record_id__delete",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{record_id}/review": {
      "patch": {
        "tags": [
          "Contracts"
        ],
        "summary": "Review Contract",
        "description": "Mark a contract as reviewed.",
        "operationId": "review_contract_contracts__record_id__review_patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractReviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{record_id}/fields": {
      "patch": {
        "tags": [
          "Contracts"
        ],
        "summary": "Update Contract Field",
        "description": "Update a single field and log the correction for ML training.\n\nThis endpoint:\n1. Updates the specified field in Airtable\n2. Logs the correction to a separate Corrections table if the value changed\n3. Returns success status and whether a correction was logged\n\nCorrections are used to build a training dataset for improving AI extraction.",
        "operationId": "update_contract_field_contracts__record_id__fields_patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid field name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "List Contracts",
        "description": "List contracts with optional status filter.",
        "operationId": "list_contracts_contracts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: 'under_review' or 'reviewed'",
              "title": "Status"
            },
            "description": "Filter by status: 'under_review' or 'reviewed'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of records to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_contract_contracts_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "PDF contract file to upload"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_contract_contracts_upload_post"
      },
      "ContractDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Deleted contract record ID"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ContractDeleteResponse",
        "description": "Response from DELETE /contracts/{id}."
      },
...
      "ContractRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Airtable record ID"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields",
            "description": "Record fields"
          },
          "created_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fields"
        ],
        "title": "ContractRecord",
        "description": "A contract record from Airtable."
      },
      "ContractReviewRequest": {
        "properties": {
          "reviewed": {
            "type": "boolean",
            "title": "Reviewed",
            "default": true
          }
        },
        "type": "object",
        "title": "ContractReviewRequest",
        "description": "Request body for PATCH /contracts/{id}/review."
      },
      "ContractReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "ContractReviewResponse",
        "description": "Response from PATCH /contracts/{id}/review."
      },
      "ContractUploadResponse": {
        "properties": {
          "contract_id": {
            "type": "string",
            "title": "Contract Id",
            "description": "Airtable record ID"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original PDF filename"
          },
          "extraction": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extraction",
            "description": "Extracted metadata fields"
          },
          "computed_dates": {
            "additionalProperties": true,
            "type": "object",
            "title": "Computed Dates",
            "description": "Computed date values"
          },
          "status": {
            "type": "string",
            "enum": [
              "under_review",
              "reviewed"
            ],
            "title": "Status",
            "default": "under_review"
          },
          "airtable_url": {
            "type": "string",
            "title": "Airtable Url",
            "description": "Direct link to Airtable record"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage",
            "description": "Token usage stats"
          }
        },
        "type": "object",
        "required": [
          "contract_id",
          "filename",
          "extraction",
          "computed_dates",
          "airtable_url"
        ],
        "title": "ContractUploadResponse",
        "description": "Response from POST /contracts/upload."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Error response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}